+++ sap_nw/startup.sh
@@ -15,18 +15,24 @@
 #
 # Description:  Google Cloud Platform - SAP Deployment Functions
 # Build Date:   BUILD.SH_DATE
+# Build Hash:   BUILD.HASH
 # ------------------------------------------------------------------------
 
 ## Check to see if a custom script path was provieded by the template
 if [[ "${1}" ]]; then
   readonly DEPLOY_URL="${1}"
 else
-  readonly DEPLOY_URL="https://storage.googleapis.com/BUILD.SH_URL"
+  readonly DEPLOY_URL="BUILD.SH_URL"
 fi
 
-## Import includes
-source /dev/stdin <<< "$(curl -s ${DEPLOY_URL}/lib/sap_lib_main.sh)"
-source /dev/stdin <<< "$(curl -s ${DEPLOY_URL}/lib/sap_lib_nw.sh)"
+##########################################################################
+## Start includes
+##########################################################################
+SAP_LIB_MAIN_SH
+SAP_LIB_NW_SH
+##########################################################################
+## End includes
+##########################################################################
 
 ### Base GCP and OS Configuration
 main::get_os_version